Separate variational pybind code into header and source file
